Updating some stuff (SetOwningCountry is really SetOwningHouse !)
git-svn-id: svn://svn.renegadeprojects.com/yrpp/trunk@144 bbc44d77-a14d-0410-bbfb-905795908d03

+++ Drawing.h
@@ -79,7 +79,7 @@ public:
 	static ColorStruct WordColor(WORD bits)
 		{ ColorStruct color;
 			color.R = ((bits & 0xF800) >> 11) << 3;
-			color.G = ((bits & 0x07E0) >> 5) << 2;
+			color.G = (BYTE)((bits & 0x07E0) >> 5) << 2; // msvc stupid warning
 			color.B = (bits & 0x001F) << 3;
 		return color; }
 };
